php: update helloworld and route guide after beta release

+++ test/core/httpcli/httpcli_test.c
@@ -134,7 +134,7 @@ int main(int argc, char **argv) {
 
   /* figure out where we are */
   if (lslash) {
-    memcpy(root, me, (size_t)(lslash - me));
+    memcpy(root, me, lslash - me);
     root[lslash - me] = 0;
   } else {
     strcpy(root, ".");
